examples/1-strctured-output/pyproject.toml
[project]
name = "1-strctured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.25.1",
    "openai>=1.105.0",
    "python-dotenv>=1.1.1",
]
